feat: Returns refresh fragment on attachments
It adds the refresh fragment to the attachment URL. It can be used to
identify the attachment object when included in other fields.

This commit also includes a refactor to move funciotns from
mdrender.extensions.refresh_attachment to the common
projects.attachments.services module.
