+++ taiga/projects/attachments/services.py
@@ -13,10 +13,16 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+from urllib.parse import parse_qs, urldefrag
+
 from django.apps import apps
 from django.conf import settings
 
 from taiga.base.utils.thumbnails import get_thumbnail_url, get_thumbnail
+
+# Refresh feature
+
+REFRESH_PARAM = "_taiga-refresh"
 
 
 def get_attachment_by_id(project_id, attachment_id):
@@ -31,6 +37,40 @@ def get_attachment_by_id(project_id, attachment_id):
 
     return obj
 
+
+def generate_refresh_fragment(attachment, type_=""):
+    if not attachment:
+        return ''
+    type_ = attachment.content_type.name if not type_ else type_
+    return "{}={}:{}".format(REFRESH_PARAM, type_, attachment.id)
+
+
+def extract_refresh_id(url):
+    if not url:
+        return False, False
+    _, frag = urldefrag(url)
+    if not frag:
+        return False, False
+    qs = parse_qs(frag)
+    if not qs:
+        return False, False
+    ref = qs.get(REFRESH_PARAM, False)
+    if not ref:
+        return False, False
+    type_, _, id_ = ref[0].partition(":")
+    try:
+        return type_, int(id_)
+    except ValueError:
+        return False, False
+
+
+def url_is_an_attachment(url: str, base=None) -> "Union[str, None]":
+    if not url:
+        return None
+    return url if url.startswith(base or settings.MEDIA_URL) else None
+
+
+# Thumbnail services
 
 def get_timeline_image_thumbnail_name(attachment):
     if attachment.attached_file:
@@ -49,9 +89,3 @@ def get_attachment_image_preview_url(attachment):
     if attachment.attached_file:
         return get_thumbnail_url(attachment.attached_file, settings.THN_ATTACHMENT_PREVIEW)
     return None
-
-
-def url_is_an_attachment(url: str, base=None) -> "Union[str, None]":
-    if not url:
-        return None
-    return url if url.startswith(base or settings.MEDIA_URL) else None
