+++ tests/unit/test_attachments_services.py
@@ -1,3 +1,5 @@
+from unittest import mock
+
 import pytest
 
 from taiga.projects.attachments import services
@@ -31,3 +33,29 @@ def test_get_attachment_by_id(django_assert_num_queries):
 ])
 def test_url_is_an_attachment(url, expected):
     assert services.url_is_an_attachment(url, base="http://media.example.com/a/") == expected
+
+
+@pytest.mark.django_db(transaction=True)
+def test_generate_refresh_fragment():
+    att = f.IssueAttachmentFactory()
+    frag = services.generate_refresh_fragment(att)
+    assert "{}=issue:{}".format(services.REFRESH_PARAM, att.id) == frag
+
+
+@pytest.mark.parametrize("url, expected", [
+    ("http://static.test/a/file.png", (False, False)),
+    ("http://static.test/a/file.png?token=x", (False, False)),
+    ("http://static.test/a/file.png?token=x#spurious", (False, False)),
+    ("http://static.test/a/file.png?token=x#" + services.REFRESH_PARAM, (False, False)),
+    ("http://static.test/a/file.png?token=x#" + services.REFRESH_PARAM + "=xxx", (False, False)),
+    ("http://static.test/a/file.png?token=x#" + services.REFRESH_PARAM + "=us:42", ("us", 42)),
+])
+def test_render_attachment_extract_refresh_id(url, expected):
+    assert services.extract_refresh_id(url) == expected
+
+
+@pytest.mark.parametrize("attachment, expected", [
+    (mock.MagicMock(id=42), services.REFRESH_PARAM + "=us:42"),
+])
+def test_generate_refresh_fragment_with_type(attachment, expected):
+    assert services.generate_refresh_fragment(attachment, "us") == expected
